move profile sub pages to pages folder
